pyproject.toml
[tool.poetry]
name = "internet-shop-frontend"
version = "0.1.0"
description = ""
authors = ["blessed <ytminicraftyt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
requests = "^2.32.3"
flask-login = "^0.6.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
